+++ tests/unit/mocks/config.js
@@ -1,3 +1,5 @@
+var _ = require('lodash');
+
 module.exports = Config;
 
 
@@ -6,11 +8,23 @@ module.exports = Config;
  *
  */
 function Config(data) {
+    var store = data || {};
+
     this.stores = {
         literal: {
-            store: data || {}
+            store: store
         }
     };
+
+    this.nconf = {
+        stores: {
+            defaults: {
+                store: store
+            }
+        }
+    };
+
+    this._store = store;
 }
 
 
@@ -19,9 +33,34 @@ function Config(data) {
  *
  * @param {String} key
  *
- * @return {undefined}
+ * @return {mixed}
  */
 Config.prototype.get = function(key) {
-    return this.stores.literal.store[key];
+    var keys = (key || '').split(':');
+    return _.get(this._store, keys);
 };
 
+/**
+ * getOrFail
+ *
+ * @param {String} key
+ *
+ * @return {mixed}
+ */
+Config.prototype.getOrFail = function(key) {
+    var keys = (key || '').split(':');
+    var val = _.get(this._store, keys);
+    if (val === undefined) {
+        throw new Error(`Cant find config value of "${key}"`);
+    }
+    return val;
+};
+
+/**
+ * @param {Object} data
+ *
+ * @return {Config}
+ */
+Config.prototype.createLiteralProvider = function(data) {
+    return new Config(data);
+};